@@ -3,15 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@react-pdf/renderer": "^2.2.0",
+    "@react-pdf/renderer": "^4.2.2",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
-    "react": "^18.3.1",
+    "core-js": "^3.40.0",
+    "react": "^19.0.0",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "react-github-calendar": "^3.2.2",
     "react-icons": "^4.8.0",
     "react-parallax-tilt": "^1.7.42",
